+++ backend/app/settings.py
@@ -13,7 +13,7 @@ https://docs.djangoproject.com/en/3.1/ref/settings/
 import os
 from pathlib import Path
 from environs import Env
-from garpix_notify.settings import *
+from garpix_notify.settings import *  # noqa
 
 env = Env()
 env.read_env()
@@ -163,7 +163,6 @@ EXAMPLE_EVENT_1 = 3
 EXAMPLE_EVENT_2 = 4
 TEST_EVENT_2 = 5
 
-
 NOTIFY_EVENTS = {
     REGISTRATION_EVENT: {
         'title': 'Register',
@@ -177,10 +176,9 @@ NOTIFY_EVENTS = {
     EXAMPLE_EVENT_2: {
         'title': 'Example 2',
     },
-    TEST_EVENT_2:
-        {
-            'title': 'TestEvent',
-        },
+    TEST_EVENT_2: {
+        'title': 'TestEvent',
+    },
 }
 
 CHOICES_NOTIFY_EVENT = [(k, v['title']) for k, v in NOTIFY_EVENTS.items()]
@@ -230,3 +228,5 @@ CHANNEL_LAYERS = {
         },
     },
 }
+
+GARPIX_NOTIFY_MIXIN = 'garpix_notify.mixins.notify_mixin.NotifyMixin'