+++ backend/app/views.py
@@ -5,10 +5,17 @@ from garpix_notify.models import Notify
 from django.conf import settings
 
 
+# Перед началом создайте шаблон для регистрации
+
+
 def example_send_notify(request):
-    Notify.send(settings.REGISTRATION_EVENT, {
-        'confirmation_code': 'abcdef12345',
-        'message': 'Привет, тут твое сообщение'
-    }, email='aleksey@garpix.com', send_at=(datetime.datetime.now() + datetime.timedelta(days=1))
-                )
+    Notify.send(
+        settings.REGISTRATION_EVENT,
+        {
+            'confirmation_code': 'abcdef12345',
+            'message': 'Привет, тут твое сообщение'
+        },
+        email='example@garpix.com',
+        send_at=(datetime.datetime.now() + datetime.timedelta(days=1))
+    )
     return render(request, 'example.html')